requirements.txt
Django==4.2.7
djangorestframework==3.14.0
openai-whisper==20231117
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
groq==0.4.1
python-dotenv==1.0.0
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
